Started adding footprints, changed values.
Added footprints for most components. need to check in uni which sizes are available and what will need to be ordered from mouser.
Changed line in/out to TRS as this will be more reliable and added 0 ohm resistor options to bypass mono input and use line in/out as stereo.