Fixed power connections on CODEC
So the codec board regulates its own AVDD which we had named but disconnected. In order for this regulation to happen the EN pin must be pulled high (not low as it says in the documentation).
SPKVDD was also disconnected as by default this is taken from VIN (5V) anyway, this means the CODEC boards don't need to be modified anymore.
Still need to test these on the board if time allows.